runtime-rs: align ch API models with CLH v51.1
Align the local runtime-rs API models with the shared field definitions
in Cloud Hypervisor v51.1.

Update CPU configuration, topology, affinity, NUMA CPU ID, VM resize,
and vsock CID types. This also preserves Kata's supported 256-vCPU
value during VM creation and resize. Remove the unused
MAX_NUM_PCI_SEGMENTS constant and add regression coverage.

Fixes: #13449

+++ src/runtime-rs/crates/hypervisor/src/ch/inner_hypervisor.rs
@@ -935,7 +935,7 @@ impl CloudHypervisorInner {
         }
 
         let vmresize = VmResize {
-            desired_vcpus: Some(new_vcpus as u8),
+            desired_vcpus: Some(new_vcpus),
             ..Default::default()
         };
